fix: rename FileLesson to keep consistency
