+++ src/tutorials/Mutable-File-System/05.vue
@@ -11,7 +11,7 @@
 </template>
 
 <script>
-import FileLesson from '../../components/File-Lesson.vue'
+import FileLesson from '../../components/FileLesson.vue'
 import text from './05.md'
 import exercise from './05-exercise.md'
