+++ src/tutorials/Mutable-File-System/07.vue
@@ -12,7 +12,7 @@
 </template>
 
 <script>
-import FileLesson from '../../components/File-Lesson.vue'
+import FileLesson from '../../components/FileLesson.vue'
 import text from './07.md'
 import exercise from './07-exercise.md'
