+++ src/tutorials/Mutable-File-System/11.vue
@@ -12,7 +12,7 @@
 </template>
 
 <script>
-import FileLesson from '../../components/File-Lesson'
+import FileLesson from '../../components/FileLesson'
 import text from './11.md'
 import exercise from './11-exercise.md'
